# Added reflection emit member accessor