… src/System.Text.Json/src/System/Text/Json/Serialization/ClassMaterializer.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/ReflectionEmitMaterializer.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/ReflectionMaterializer.cs