This is prerequisite for #36505 and opened as a separate PR because it can be viewed as an optimization of existing code. The feature will come after merging this one.

/cc @ahsonkhan @steveharter