@YohDeadfall 2019-06-06
I have a question about an empty line at the end of file. What style is preferred? In some files I see no empty lines, but in other they are.