@steveharter 2019-06-07
>> I have a question about an empty line at the end of file. What style is preferred? In some files I see no empty lines, but in other they are.

Just make sure there's a linefeed\CR on the last line - no extra empty line is desired AFAIK.