@steveharter 2019-06-14
The tests are failing --  IL causing InvalidProgramException?
```
System.Text.Json.Serialization.Tests.ExtensionDataTests.ExtensionPropertyObjectValue_RoundTrip

Unhandled Exception of Type System.InvalidProgramException
Message :
System.InvalidProgramException : Common Language Runtime detected an invalid program.
Stack Trace :
   at get_Prop(Object )
   at System.Text.Json.JsonPropertyInfoNotNullable`3.Write(WriteStackFrame& current, Utf8JsonWriter writer)
   at System.Text.Json.JsonSerializer.HandleObject(JsonSerializerOptions options, Utf8JsonWriter writer, WriteStack& state)
   at System.Text.Json.JsonSerializer.WriteObject(JsonSerializerOptions options, Utf8JsonWriter writer, WriteStack& state)
   at System.Text.Json.JsonSerializer.Write(Utf8JsonWriter writer, Int32 flushThreshold, JsonSerializerOptions options, WriteStack& state)
   at System.Text.Json.JsonSerializer.WriteCore(PooledByteBufferWriter output, Object value, Type type, JsonSerializerOptions options)
   at System.Text.Json.JsonSerializer.WriteCoreString(Object value, Type type, JsonSerializerOptions options)
   at System.Text.Json.JsonSerializer.ToStringInternal(Object value, Type type, JsonSerializerOptions options)
   at System.Text.Json.JsonSerializer.ToString[TValue](TValue value, JsonSerializerOptions options)
   at System.Text.Json.Serialization.Tests.ExtensionDataTests.ExtensionPropertyObjectValue_RoundTrip()
```