@YohDeadfall 2019-06-14
> The tests are failing -- IL causing InvalidProgramException?

How can I view the previous pipeline run results? It seems that it was more successful.