@steveharter 2019-06-15
> How can I view the previous pipeline run results? It seems that it was more successful, but may be wrong.

Usually they are available in the commit history, but I am not sure how long they remain there.