@YohDeadfall 2019-06-15
Now there is only one failed test and it's unrelated to this PR:

```
System.ComponentModel.Win32Exception : Access is denied.
   at System.Diagnostics.EventLogInternal.InternalWriteEvent(UInt32 eventID, UInt16 category, EventLogEntryType type, String[] strings, Byte[] rawData, String currentMachineName) in /_/src/System.Diagnostics.EventLog/src/System/Diagnostics/EventLogInternal.cs:line 1441
   at System.Diagnostics.EventLogInternal.WriteEvent(EventInstance instance, Byte[] data, Object[] values) in /_/src/System.Diagnostics.EventLog/src/System/Diagnostics/EventLogInternal.cs:line 1382
   at System.Diagnostics.EventLog.WriteEvent(EventInstance instance, Byte[] data, Object[] values) in /_/src/System.Diagnostics.EventLog/src/System/Diagnostics/EventLog.cs:line 1003
   at System.Diagnostics.EventLog.WriteEvent(EventInstance instance, Object[] values) in /_/src/System.Diagnostics.EventLog/src/System/Diagnostics/EventLog.cs:line 998
   at System.Diagnostics.EventLogTraceListener.TraceEvent(TraceEventCache eventCache, String source, TraceEventType severity, Int32 id, String message) in /_/src/System.Diagnostics.EventLog/src/System/Diagnostics/EventLogTraceListener.cs:line 116
   at System.Diagnostics.Tests.EventLogTraceListenerTests.TraceEventTest(TraceEventType eventType, EventLogEntryType expectedType, Int32 id, Int32 expectedId) in /_/src/System.Diagnostics.EventLog/tests/EventLogTraceListenerTests.cs:line 131
```