@steveharter 2019-06-15
@YohDeadfall I think we need to pause for a couple days until we can review this and ensure boxing is not affecting perf.

Also, with the support for [custom converters](https://github.com/dotnet/corefx/issues/36639) it is possible to write a converter for a struct, including a standard struct scenario where there is a constructor that takes the arguments and does not have setters for any properties.

What are the viable scenarios where a large struct with several properties is desired over a class (and no converter is desired)?