@YohDeadfall 2019-06-17
@steveharter My pull request introduces no boxing. It just adds code generation which does the same thing as two hops with delegates. The previous one didn't that too. The only thing that I changed is moved casting of an `object` to `TClass` into generated setters/getters.

Improving performance via removing boxing is possible and I see to ways to do that:

1. Change the frame type to be a class and store instances as linked list nodes. In that case you will have a base frame class and generic one. Because all pointers takes the same size of memory it makes sense to have a generic frame class for structures and another one for classes to store `object`s. Frames could be rented from types pools.

2. Use a simple class to store boxed values on the heap, but that generic class should be rentable as in the previous approach. The same trick we use in Npgsql to work with value types in async methods, but without pooling now (see the [type](https://github.com/npgsql/npgsql/blob/28a187fe7b2b14f504d7673c42bd8602253a3e46/src/Npgsql/TypeHandlers/CompositeHandlers/ByReference.cs#L7-L10) and a [setter](https://github.com/npgsql/npgsql/blob/28a187fe7b2b14f504d7673c42bd8602253a3e46/src/Npgsql/TypeHandlers/CompositeHandlers/CompositeMemberHandlerOfStruct.cs#L106-L112) for example).

```csharp
internal sealed class ByReference<T>
{
    public T Value;
}
```

> Also, with the support for custom converters it is possible to write a converter for a struct, including a standard struct scenario where there is a constructor that takes the arguments and does not have setters for any properties.

Recommended this way as a temporary solution to the ASP.NET Core team, but as I know they chose another simpler way. Probably you should ask @rynowak or someone else working in that area. But as a customer I wouldn't like to write a custom converter too.

> What are the viable scenarios where a large struct with several properties is desired over a class (and no converter is desired)?

Not large, but small. For example, typified identifiers, coordinates and similar cases. No one would like to store them on the heap.