@steveharter 2019-06-18
> My pull request introduces no boxing. It just adds code generation which does the same thing as two hops with delegates. The previous one didn't that too. The only thing that I changed is moved casting of an object to TClass into generated setters/getters.

The previous code on a POCO with a property did not box on get\set since the value was going through a direct delegate because both the [class type and property type are static via `Action<TClass, TProperty>`.](https://github.com/dotnet/corefx/blob/master/src/System.Text.Json/src/System/Text/Json/Serialization/MemberAccessor.cs#L18)

The new code doesn't use a static type for the class and instead now [uses 'object'] instead of `TClass`(https://github.com/dotnet/corefx/pull/38233/files#diff-16a59a58106bf15daf23e3758ef24461R115) which means it will no longer be a direct delegate.

If you can provide before\after benchmark results on a POCO object with primitives (int, DateTime, etc) then we can compare and go from there. Thanks