@steveharter 2019-06-19
@YohDeadfall changes here need to be vetted through benchmarks to ensure no regressions. This is a sensitive area where a lot of effort that went into it to make it performant.

If you can provide before\after benchmark results on a POCO object with a few primitives (int, DateTime, etc) then we can compare and go from there. Thanks

The benchmarks are at https://github.com/dotnet/performance/tree/master/src/benchmarks/micro/Serializers. I can run them for you.