@YohDeadfall 2019-06-25
``` ini

BenchmarkDotNet=v0.11.3.1003-nightly, OS=Windows 10.0.18895
Intel Xeon CPU E5-2687W 0 3.10GHz, 2 CPU, 8 logical and 8 physical cores
.NET Core SDK=3.0.100-preview7-012598
  [Host]     : .NET Core 3.0.0-preview7-27824-03 (CoreCLR 4.700.19.32302, CoreFX 4.700.19.32001), 64bit RyuJIT
  Job-DAJFDT : .NET Core ? (CoreCLR 4.700.19.32003, CoreFX 4.700.19.32201), 64bit RyuJIT

Runtime=Core  Toolchain=CoreRun  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

```

```
dotnet run --base "D:\Repos\corefxperf\before" --diff "D:\Repos\corefxperf\after" --threshold 5%
```

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Text.Json.Serialization.Tests.WriteJson<LoginViewModel>.SerializeToStream |      1.13 |          2123.43 |          2393.33 | bimodal |
| System.Text.Json.Serialization.Tests.WriteJson<BinaryData>.SerializeToStream     |      1.09 |         12348.02 |         13430.81 | several?|

No Faster results for the provided threshold = 5% and noise filter = 0.3ns.

# ReadJson
## MyEventsListerViewModel
### Before
|                   Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 1.883 ms | 0.1103 ms | 0.1270 ms | 1.906 ms | 1.609 ms | 2.082 ms |     14.9254 |           - |           - |           153.84 KB |
| DeserializeFromUtf8Bytes | 1.765 ms | 0.1280 ms | 0.1474 ms | 1.789 ms | 1.488 ms | 2.000 ms |      7.0423 |           - |           - |            74.46 KB |
|    DeserializeFromStream | 2.157 ms | 0.1368 ms | 0.1575 ms | 2.193 ms | 1.700 ms | 2.399 ms |           - |           - |           - |            74.53 KB |
### After
|                   Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 1.953 ms | 0.1005 ms | 0.1158 ms | 1.969 ms | 1.662 ms | 2.171 ms |      8.0000 |           - |           - |           154.34 KB |
| DeserializeFromUtf8Bytes | 1.751 ms | 0.1312 ms | 0.1511 ms | 1.791 ms | 1.417 ms | 1.938 ms |      6.8966 |           - |           - |            74.46 KB |
|    DeserializeFromStream | 2.126 ms | 0.1238 ms | 0.1425 ms | 2.142 ms | 1.824 ms | 2.390 ms |           - |           - |           - |            74.53 KB |


## LoginViewModel
### Before
|                   Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 2.770 us | 0.1055 us | 0.1215 us | 2.741 us | 2.593 us | 2.982 us |      0.0219 |           - |           - |               280 B |
| DeserializeFromUtf8Bytes | 2.589 us | 0.1252 us | 0.1441 us | 2.581 us | 2.228 us | 2.828 us |      0.0100 |           - |           - |               168 B |
|    DeserializeFromStream | 3.959 us | 0.1305 us | 0.1502 us | 3.939 us | 3.667 us | 4.327 us |      0.0130 |           - |           - |               240 B |
### After
|                   Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 2.608 us | 0.1647 us | 0.1830 us | 2.673 us | 2.118 us | 2.858 us |      0.0194 |           - |           - |               280 B |
| DeserializeFromUtf8Bytes | 2.578 us | 0.0873 us | 0.1005 us | 2.598 us | 2.406 us | 2.719 us |      0.0104 |           - |           - |               168 B |
|    DeserializeFromStream | 3.980 us | 0.3105 us | 0.3576 us | 4.008 us | 3.293 us | 4.562 us |      0.0145 |           - |           - |               240 B |

## Location
### Before
|                   Method |     Mean |     Error |    StdDev |   Median |      Min |       Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|----------:|----------:|---------:|---------:|----------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 7.656 us | 0.3909 us | 0.4502 us | 7.638 us | 6.516 us |  8.280 us |      0.0627 |           - |           - |               680 B |
| DeserializeFromUtf8Bytes | 6.873 us | 0.5833 us | 0.6718 us | 7.078 us | 4.969 us |  7.506 us |      0.0279 |           - |           - |               448 B |
|    DeserializeFromStream | 9.401 us | 0.7543 us | 0.8686 us | 9.679 us | 7.583 us | 10.806 us |      0.0374 |           - |           - |               520 B |
### After
|                   Method |     Mean |     Error |    StdDev |   Median |      Min |       Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|----------:|----------:|---------:|---------:|----------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 7.765 us | 0.4897 us | 0.5640 us | 7.804 us | 6.521 us |  8.576 us |      0.0645 |           - |           - |               680 B |
| DeserializeFromUtf8Bytes | 7.218 us | 0.3794 us | 0.4370 us | 7.204 us | 6.107 us |  7.782 us |      0.0251 |           - |           - |               448 B |
|    DeserializeFromStream | 9.810 us | 0.5751 us | 0.6623 us | 9.874 us | 8.186 us | 10.929 us |      0.0372 |           - |           - |               520 B |

## IndexViewModel
### Before
|                   Method |     Mean |    Error |   StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|---------:|---------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 200.6 us | 21.36 us | 24.60 us | 206.9 us | 141.1 us | 230.9 us |      3.2600 |           - |           - |            34.82 KB |
| DeserializeFromUtf8Bytes | 192.9 us | 11.36 us | 12.63 us | 195.0 us | 167.9 us | 211.0 us |      1.6393 |           - |           - |            21.91 KB |
|    DeserializeFromStream | 194.5 us | 13.81 us | 15.91 us | 196.0 us | 156.9 us | 220.5 us |      1.7316 |           - |           - |            21.98 KB |
### After
|                   Method |     Mean |    Error |   StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|---------:|---------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 190.3 us | 10.54 us | 12.14 us | 195.6 us | 167.5 us | 204.5 us |      3.0604 |           - |           - |             34.8 KB |
| DeserializeFromUtf8Bytes | 195.7 us | 10.00 us | 11.52 us | 195.2 us | 171.3 us | 211.5 us |      1.5601 |           - |           - |            21.91 KB |
|    DeserializeFromStream | 199.2 us | 12.52 us | 14.41 us | 201.3 us | 161.4 us | 222.7 us |      1.6103 |           - |           - |            21.98 KB |

## BinaryData
### Before
|                   Method |     Mean |    Error |   StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|---------:|---------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 24.30 us | 2.618 us | 3.015 us | 25.32 us | 18.25 us | 27.92 us |      0.2067 |           - |           - |             2.63 KB |
| DeserializeFromUtf8Bytes | 23.70 us | 1.575 us | 1.813 us | 23.93 us | 20.10 us | 26.66 us |      0.0818 |           - |           - |             1.05 KB |
|    DeserializeFromStream | 25.81 us | 1.419 us | 1.577 us | 25.95 us | 22.34 us | 28.30 us |      0.0903 |           - |           - |             1.12 KB |
### After
|                   Method |     Mean |    Error |   StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|---------:|---------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 26.27 us | 1.235 us | 1.423 us | 26.28 us | 22.30 us | 28.88 us |      0.2016 |           - |           - |             2.63 KB |
| DeserializeFromUtf8Bytes | 24.30 us | 1.439 us | 1.658 us | 24.57 us | 21.14 us | 27.10 us |      0.1011 |           - |           - |             1.05 KB |
|    DeserializeFromStream | 27.46 us | 1.017 us | 1.171 us | 27.30 us | 25.48 us | 30.21 us |      0.1038 |           - |           - |             1.12 KB |


# WriteJson
## MyEventsListerViewModel
### Before
|               Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 3.006 ms | 0.2388 ms | 0.2654 ms | 3.018 ms | 2.468 ms | 3.530 ms |     41.6667 |     41.6667 |     41.6667 |           392.11 KB |
| SerializeToUtf8Bytes | 2.924 ms | 0.1559 ms | 0.1600 ms | 2.958 ms | 2.495 ms | 3.095 ms |     20.8333 |           - |           - |           312.69 KB |
|    SerializeToStream | 2.888 ms | 0.2319 ms | 0.2577 ms | 2.957 ms | 2.386 ms | 3.323 ms |     20.8333 |           - |           - |           232.84 KB |
### After
|               Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 3.251 ms | 0.1841 ms | 0.1808 ms | 3.328 ms | 2.655 ms | 3.364 ms |     41.6667 |     41.6667 |     41.6667 |           392.58 KB |
| SerializeToUtf8Bytes | 2.997 ms | 0.1781 ms | 0.1980 ms | 2.982 ms | 2.680 ms | 3.398 ms |     20.8333 |           - |           - |           312.73 KB |
|    SerializeToStream | 2.908 ms | 0.1667 ms | 0.1712 ms | 2.981 ms | 2.418 ms | 3.086 ms |     20.8333 |           - |           - |           232.84 KB |

## LoginViewModel
### Before
|               Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 1.914 us | 0.1726 us | 0.1988 us | 1.985 us | 1.387 us | 2.184 us |      0.0244 |           - |           - |               336 B |
| SerializeToUtf8Bytes | 1.742 us | 0.1123 us | 0.1293 us | 1.751 us | 1.391 us | 1.956 us |      0.0214 |           - |           - |               256 B |
|    SerializeToStream | 2.077 us | 0.1623 us | 0.1869 us | 2.123 us | 1.709 us | 2.383 us |      0.0086 |           - |           - |               144 B |

### After
|               Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 1.949 us | 0.1079 us | 0.1243 us | 1.961 us | 1.703 us | 2.148 us |      0.0274 |           - |           - |               336 B |
| SerializeToUtf8Bytes | 1.860 us | 0.0996 us | 0.1147 us | 1.857 us | 1.626 us | 2.053 us |      0.0218 |           - |           - |               256 B |
|    SerializeToStream | 2.405 us | 0.0935 us | 0.1077 us | 2.393 us | 2.229 us | 2.613 us |      0.0084 |           - |           - |               144 B |

## Location
### Before
|               Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 135.3 us | 12.414 us | 14.297 us | 137.3 us | 111.3 us | 156.6 us |      2.1186 |           - |           - |             26856 B |
| SerializeToUtf8Bytes | 137.9 us |  7.082 us |  8.156 us | 137.8 us | 125.7 us | 152.7 us |      1.1468 |           - |           - |             13680 B |
|    SerializeToStream | 138.8 us | 11.805 us | 13.594 us | 141.8 us | 114.7 us | 156.9 us |           - |           - |           - |               464 B |

### After
|               Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 4.541 us | 0.2534 us | 0.2919 us | 4.557 us | 3.963 us | 5.085 us |      0.0542 |           - |           - |               584 B |
| SerializeToUtf8Bytes | 4.468 us | 0.2303 us | 0.2652 us | 4.549 us | 3.813 us | 4.845 us |      0.0336 |           - |           - |               376 B |
|    SerializeToStream | 4.863 us | 0.2772 us | 0.3192 us | 4.874 us | 4.181 us | 5.418 us |           - |           - |           - |               144 B |

## IndexViewModel
### Before
|               Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 4.523 us | 0.2303 us | 0.2464 us | 4.618 us | 4.006 us | 4.769 us |      0.0547 |           - |           - |               584 B |
| SerializeToUtf8Bytes | 4.262 us | 0.2753 us | 0.3060 us | 4.365 us | 3.379 us | 4.653 us |      0.0342 |           - |           - |               376 B |
|    SerializeToStream | 4.649 us | 0.3536 us | 0.4072 us | 4.755 us | 3.679 us | 5.253 us |           - |           - |           - |               144 B |
### After
|               Method |     Mean |     Error |    StdDev |   Median |       Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|----------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 146.1 us |  7.805 us |  8.988 us | 147.2 us | 127.08 us | 160.7 us |      2.1552 |           - |           - |             26856 B |
| SerializeToUtf8Bytes | 145.3 us |  5.916 us |  6.576 us | 145.1 us | 124.27 us | 154.9 us |      1.1792 |           - |           - |             13680 B |
|    SerializeToStream | 135.4 us | 13.636 us | 15.703 us | 140.9 us |  98.52 us | 150.6 us |           - |           - |           - |               464 B |

## BinaryData
### Before
|               Method |     Mean |     Error |    StdDev |   Median |       Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|----------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 13.11 us | 0.9075 us | 1.0451 us | 13.24 us | 10.177 us | 14.46 us |      0.2729 |           - |           - |              3368 B |
| SerializeToUtf8Bytes | 12.39 us | 0.8534 us | 0.9828 us | 12.61 us | 10.537 us | 13.88 us |      0.1511 |           - |           - |              1768 B |
|    SerializeToStream | 11.76 us | 1.2537 us | 1.4437 us | 12.35 us |  8.092 us | 13.57 us |           - |           - |           - |               144 B |
### After
|               Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    SerializeToString | 13.61 us | 0.6727 us | 0.7746 us | 13.75 us | 11.74 us | 14.62 us |      0.2732 |           - |           - |              3368 B |
| SerializeToUtf8Bytes | 12.86 us | 0.6981 us | 0.8039 us | 12.88 us | 10.43 us | 14.16 us |      0.1303 |           - |           - |              1768 B |
|    SerializeToStream | 13.14 us | 0.9132 us | 1.0516 us | 13.43 us | 10.55 us | 14.47 us |           - |           - |           - |               144 B |